Refatoracao dos pacotes de endereco, medico e paciente para dentro do pacote domain
