+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/address/Address.java
@@ -1,4 +1,4 @@
-package br.com.feliperochasi.med.voll.api.address;
+package br.com.feliperochasi.med.voll.api.domain.address;
 
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
