+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/medic/MedicRepository.java
@@ -1,10 +1,8 @@
-package br.com.feliperochasi.med.voll.api.medic;
+package br.com.feliperochasi.med.voll.api.domain.medic;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface MedicRepository extends JpaRepository<Medic, Long> {
     Page<Medic> findAllByActiveTrue(Pageable p);
