+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/medic/RegisterDataMedic.java
@@ -1,25 +1,26 @@
-package br.com.feliperochasi.med.voll.api.patient;
+package br.com.feliperochasi.med.voll.api.domain.medic;
 
-import br.com.feliperochasi.med.voll.api.address.Address;
-import br.com.feliperochasi.med.voll.api.address.DataAddress;
+import br.com.feliperochasi.med.voll.api.domain.address.DataAddress;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
-public record RegisterDataPatient(
+public record RegisterDataMedic(
         @NotBlank
         String nome,
         @NotBlank
         @Email
         String email,
         @NotBlank
-        String telefone,
-        @NotBlank
-        @Pattern(regexp = "\\d{11}")
-        String cpf,
-        @Valid
+        @Pattern(regexp = "\\d{4,6}")
+        String crm,
         @NotNull
+        Specialised especialidade,
+        @NotBlank
+        String telefone,
+        @NotNull
+        @Valid
         DataAddress endereco) {
 }
