+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/medic/Specialised.java
@@ -1,4 +1,4 @@
-package br.com.feliperochasi.med.voll.api.medic;
+package br.com.feliperochasi.med.voll.api.domain.medic;
 
 public enum Specialised {
     ORTOPEDIA,
