+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/medic/UpdateDataMedic.java
@@ -1,6 +1,6 @@
-package br.com.feliperochasi.med.voll.api.medic;
+package br.com.feliperochasi.med.voll.api.domain.medic;
 
-import br.com.feliperochasi.med.voll.api.address.DataAddress;
+import br.com.feliperochasi.med.voll.api.domain.address.DataAddress;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateDataMedic(
