+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/patient/Patient.java
@@ -1,6 +1,6 @@
-package br.com.feliperochasi.med.voll.api.patient;
+package br.com.feliperochasi.med.voll.api.domain.patient;
 
-import br.com.feliperochasi.med.voll.api.address.Address;
+import br.com.feliperochasi.med.voll.api.domain.address.Address;
 import jakarta.persistence.*;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
