+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/patient/RegisterDataPatient.java
@@ -1,26 +1,24 @@
-package br.com.feliperochasi.med.voll.api.medic;
+package br.com.feliperochasi.med.voll.api.domain.patient;
 
-import br.com.feliperochasi.med.voll.api.address.DataAddress;
+import br.com.feliperochasi.med.voll.api.domain.address.DataAddress;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
-public record RegisterDataMedic(
+public record RegisterDataPatient(
         @NotBlank
         String nome,
         @NotBlank
         @Email
         String email,
         @NotBlank
-        @Pattern(regexp = "\\d{4,6}")
-        String crm,
-        @NotNull
-        Specialised especialidade,
-        @NotBlank
         String telefone,
-        @NotNull
+        @NotBlank
+        @Pattern(regexp = "\\d{11}")
+        String cpf,
         @Valid
+        @NotNull
         DataAddress endereco) {
 }
