+++ src/main/java/br/com/feliperochasi/med/voll/api/domain/patient/UpdateDataPatient.java
@@ -1,6 +1,6 @@
-package br.com.feliperochasi.med.voll.api.patient;
+package br.com.feliperochasi.med.voll.api.domain.patient;
 
-import br.com.feliperochasi.med.voll.api.address.DataAddress;
+import br.com.feliperochasi.med.voll.api.domain.address.DataAddress;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateDataPatient(
